Move save command into its own module for consistency
