+++ cmd/root.go
@@ -5,6 +5,7 @@ import (
 	"os"
 
 	"github.com/sector-f/jhmod/cmd/nvccmd"
+	"github.com/sector-f/jhmod/cmd/savecmd"
 	"github.com/spf13/cobra"
 )
 
@@ -16,7 +17,7 @@ var rootCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(nvccmd.Cmd())
-	rootCmd.AddCommand(saveCmd)
+	rootCmd.AddCommand(savecmd.Cmd())
 }
 
 func Execute() {
